services/scraper/pyproject.toml
[project]
name = "cyprus-wikimapia-scraper"
version = "1.0.0"
description = "Scrapes geodata from Wikimapia for Cyprus island and produces GeoJSON"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "shapely>=2.0",
    "geojson>=3.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scraper"]
